[839] Scope terms under publish
