+++ app/views/publish/terms/edit.html.erb
@@ -4,7 +4,7 @@
   <div class="govuk-grid-column-two-thirds">
     <%= form_with(
       model: @accept_terms_form,
-      url: accept_terms_path,
+      url: publish_accept_terms_path,
       method: :patch,
     ) do |f| %>
