+++ config/locales/en.yml
@@ -400,7 +400,7 @@ en:
           attributes:
             confirm_course_code:
               invalid_code: "Enter the course code %{course_code} to delete this course"
-        publish/accept_terms_form:
+        publish/interruption/accept_terms_form:
           attributes:
             terms_accepted:
               accepted: "You must accept the terms and conditions to continue"
